protocol: move protocols into submodule
